Additional cleanup and reorganization. Adding a failing test captured from conformance tool ALG_SIGN_SECP521R1_ECDSA_SHA512_RAW.
