+++ fido2-net-lib.Test/attestationOptionsPacked512.json
@@ -1,0 +1,1 @@
+{"rp":{"id":"localhost","name":"Fido2 test"},"user":{"name":"qhpF3If1h4ANVCNzl08m","id":"qhpF3If1h4ANVCNzl08m","displayName":"Shala Dull"},"challenge":"QPA-FrCSwg-qHhzYvRIdnA","pubKeyCredParams":[{"type":"public-key","alg":-7},{"type":"public-key","alg":-257},{"type":"public-key","alg":-37},{"type":"public-key","alg":-35},{"type":"public-key","alg":-258},{"type":"public-key","alg":-38},{"type":"public-key","alg":-36},{"type":"public-key","alg":-259},{"type":"public-key","alg":-39},{"type":"public-key","alg":-65535}],"timeout":60000,"attestation":"direct","authenticatorSelection":null,"excludeCredentials":[],"status":"ok","errorMessage":""}
