+++ fido2-net-lib.Test/fido2-net-lib.Test.csproj
@@ -15,9 +15,11 @@
     <None Remove="attestationNoneResponse.json" />
     <None Remove="attestationOptionsNone.json" />
     <None Remove="attestationOptionsPacked.json" />
+    <None Remove="attestationOptionsPacked512.json" />
     <None Remove="attestationOptionsU2F.json" />
     <None Remove="attestationResultsNone.json" />
     <None Remove="attestationResultsPacked.json" />
+    <None Remove="attestationResultsPacked512.json" />
     <None Remove="attestationResultsU2F.json" />
     <None Remove="attestationTPMSHA1Options.json" />
     <None Remove="attestationTPMSHA1Response.json" />
@@ -59,6 +61,9 @@
     <Content Include="attestationOptionsPacked.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+    <Content Include="attestationOptionsPacked512.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
     <Content Include="attestationOptionsU2F.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
@@ -66,6 +71,9 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="attestationResultsPacked.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="attestationResultsPacked512.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="attestationResultsU2F.json">
