+++ fido2-net-lib/AttestationFormat/AttestationFormat.cs
@@ -1,6 +1,8 @@
 ﻿using PeterO.Cbor;
 using Fido2NetLib.Objects;
 using System;
+using System.Security.Cryptography.X509Certificates;
+using System.Linq;
 
 namespace Fido2NetLib.AttestationFormat
 {
@@ -31,7 +33,38 @@ namespace Fido2NetLib.AttestationFormat
                 return data;
             }
         }
-
+        internal static byte[] AaguidFromAttnCertExts(X509ExtensionCollection exts)
+        {
+            byte[] aaguid = null;
+            foreach (var ext in exts)
+            {
+                if (ext.Oid.Value.Equals("1.3.6.1.4.1.45724.1.1.4")) // id-fido-gen-ce-aaguid
+                {
+                    aaguid = new byte[16];
+                    var ms = new System.IO.MemoryStream(ext.RawData.ToArray());
+                    // OCTET STRING
+                    if (0x4 != ms.ReadByte()) throw new Fido2VerificationException("Expected octet string value");
+                    // AAGUID
+                    if (0x10 != ms.ReadByte()) throw new Fido2VerificationException("Unexpected length for aaguid");
+                    ms.Read(aaguid, 0, 0x10);
+                    //The extension MUST NOT be marked as critical
+                    if (true == ext.Critical) throw new Fido2VerificationException("extension MUST NOT be marked as critical");
+                }
+            }
+            return aaguid;
+        }
+        internal static bool IsAttnCertCACert(X509ExtensionCollection exts)
+        {
+            foreach (var ext in exts)
+            {
+                if (ext.Oid.FriendlyName == "Basic Constraints")
+                {
+                    var baseExt = (X509BasicConstraintsExtension)ext;
+                    return baseExt.CertificateAuthority;
+                }
+            }
+            return true;
+        }
         public abstract AttestationFormatVerificationResult Verify();
     }
 }
